Externalize WS server
