+++ Polyhedron/demo/Polyhedron/MainWindow.cpp
@@ -77,12 +77,6 @@
 #include "Color_map.h"
 
 
-#ifdef CGAL_USE_WEBSOCKETS
-#include <QWebSocketServer>
-#include <QWebSocket>
-#include <QNetworkInterface>
-#endif
-
 using namespace CGAL::Three;
 QScriptValue
 myScene_itemToScriptValue(QScriptEngine *engine,
@@ -3681,90 +3675,3 @@ void MainWindow::on_actionLoad_a_Scene_from_a_Script_File_triggered()
     tmp_file.remove();
   }
 }
-
-#ifdef CGAL_USE_WEBSOCKETS
-void MainWindow::on_action_Start_a_Session_triggered()
-{
-  QAction * action= findChild<QAction*>("action_Start_a_Session");
-  static EchoServer *server =nullptr;
-  if(action->isChecked()){
-     server = new EchoServer(1234);
-    QObject::connect(server, &EchoServer::closed, server,&EchoServer::deleteLater);
-  }
-  else
-  {
-    server->deleteLater();
-  }
-}
-
-EchoServer::EchoServer(quint16 port) :
-    QObject(CGAL::Three::Three::mainWindow()),
-    m_pWebSocketServer(new QWebSocketServer(QStringLiteral("Echo Server"),
-                                            QWebSocketServer::NonSecureMode, this))
-{
-    if (m_pWebSocketServer->listen(QHostAddress::Any, port)) {
-        connect(m_pWebSocketServer, &QWebSocketServer::newConnection,
-                this, &EchoServer::onNewConnection);
-        connect(m_pWebSocketServer, &QWebSocketServer::closed, this, &EchoServer::closed);
-    }
-    QHostAddress local_host("0.0.0.0");
-
-    //to avoid printing 127.0.0.1. Not realy sure it won't ever print the external ipv4 though.
-    const QHostAddress &localhost = QHostAddress(QHostAddress::LocalHost);
-    for (const QHostAddress &address: QNetworkInterface::allAddresses()) {
-      if (address.protocol() == QAbstractSocket::IPv4Protocol && address != localhost)
-      {
-        local_host= address;
-        break;
-      }
-    }
-    QMessageBox mb(QMessageBox::NoIcon, "WS Server",
-                   tr("WebSockets Server started.\nEnter the following address in\nyour Network Preferences to be able to join it :\n"
-                      "ws://%1:%2").arg(local_host.toString()).arg(port), QMessageBox::Ok, CGAL::Three::Three::mainWindow());
-    mb.setTextInteractionFlags(Qt::TextSelectableByMouse);
-    mb.exec();
-}
-
-EchoServer::~EchoServer()
-{
-    m_pWebSocketServer->close();
-    qDeleteAll(m_clients.begin(), m_clients.end());
-}
-
-void EchoServer::onNewConnection()
-{
-    QWebSocket *pSocket = m_pWebSocketServer->nextPendingConnection();
-
-    connect(pSocket, &QWebSocket::textMessageReceived, this, &EchoServer::processTextMessage);
-    connect(pSocket, &QWebSocket::binaryMessageReceived, this, &EchoServer::processBinaryMessage);
-    connect(pSocket, &QWebSocket::disconnected, this, &EchoServer::socketDisconnected);
-
-    m_clients << pSocket;
-}
-
-void EchoServer::processTextMessage(QString message)
-{
-    QWebSocket *pClient = qobject_cast<QWebSocket *>(sender());
-    for(auto *client : m_clients) {
-      if(client != pClient)
-        client->sendTextMessage(message);
-    }
-}
-
-void EchoServer::processBinaryMessage(QByteArray message)
-{
-    QWebSocket *pClient = qobject_cast<QWebSocket *>(sender());
-    if (pClient) {
-        pClient->sendBinaryMessage(message);
-    }
-}
-
-void EchoServer::socketDisconnected()
-{
-    QWebSocket *pClient = qobject_cast<QWebSocket *>(sender());
-    if (pClient) {
-        m_clients.removeAll(pClient);
-        pClient->deleteLater();
-    }
-}
-#endif
